Animal Detection/requirements.txt
Flask==2.1.0
tensorflow==2.6.0
keras==2.6.0
Pillow==8.3.1
numpy==1.19.5
